Move allowance checks into shared hook.
